combine consistent region tests in one test module, #119
