+++ test/test_objectstorage_toolkit.py
@@ -447,81 +447,6 @@ class TestDistributed(unittest.TestCase):
 
     # -------------------
 
-    # CONSISTENT REGION: ObjectStorageSource (IAM), static name, text file (read 100 lines of 1 MB line size), no crash (no reset)
-    @unittest.skipIf(th.iam_credentials() == False, "Missing "+th.COS_IAM_CREDENTIALS()+" environment variable.")
-    def test_read_object_consistent_region_static_name_iam(self):
-        th.generate_large_text_file("input.txt")
-        s3.uploadObject(self.s3_client_iam, self.bucket_name_iam, "input.txt", "input.txt")
-        # periodic
-        self._build_launch_validate("test_read_object_consistent_region_static_name_periodic_iam", "com.ibm.streamsx.objectstorage.test::ReadTestConsistentRegionPeriodicStaticNameIAMComp", {'IAMApiKey':self.iam_api_key, 'IAMServiceInstanceId':self.service_instance_id, 'objectStorageURI':self.uri_cos}, 1, 'feature/consistent.region.test', True, 90)
-        # operatorDriven
-        self._build_launch_validate("test_read_object_consistent_region_static_name_operatorDriven_iam", "com.ibm.streamsx.objectstorage.test::ReadTestConsistentRegionOperatorDrivenStaticNameIAMComp", {'IAMApiKey':self.iam_api_key, 'IAMServiceInstanceId':self.service_instance_id, 'objectStorageURI':self.uri_cos}, 1, 'feature/consistent.region.test', True, 90)
-
-    # CONSISTENT REGION: S3ObjectStorageSource, static name, text file (read 100 lines of 1 MB line size), no crash (no reset)
-    @unittest.skipIf(th.cos_credentials() == False, "Missing "+th.COS_CREDENTIALS()+" environment variable.")
-    def test_read_object_consistent_region_static_name(self):
-        th.generate_large_text_file("input.txt")
-        s3.uploadObject(self.s3_client, self.bucket_name, "input.txt", "input.txt")
-        # periodic
-        self._build_launch_validate("test_read_object_consistent_region_static_name_periodic", "com.ibm.streamsx.objectstorage.test::ReadTestConsistentRegionPeriodicStaticNameComp", {'accessKeyID':self.access_key, 'secretAccessKey':self.secret_access_key, 'bucket':self.bucket_name}, 1, 'feature/consistent.region.test', True, 90)
-        # operatorDriven
-        self._build_launch_validate("test_read_object_consistent_region_static_name_operatorDriven", "com.ibm.streamsx.objectstorage.test::ReadTestConsistentRegionOperatorDrivenStaticNameComp", {'accessKeyID':self.access_key, 'secretAccessKey':self.secret_access_key, 'bucket':self.bucket_name}, 1, 'feature/consistent.region.test', True, 90)
-
-    # -------------------
-
-    # CONSISTENT REGION: ObjectStorageSource (IAM), static name, binary file (read 100 blocks of 1 MB block size), no crash (no reset)
-    @unittest.skipIf(th.iam_credentials() == False, "Missing "+th.COS_IAM_CREDENTIALS()+" environment variable.")
-    def test_read_object_consistent_region_static_name_binary_iam(self):
-        th.generate_large_bin_file("input.bin") # 100 MB
-        s3.uploadObject(self.s3_client_iam, self.bucket_name_iam, "input.bin", "input.bin")
-        # periodic
-        self._build_launch_validate("test_read_object_consistent_region_static_name_binary_periodic_iam", "com.ibm.streamsx.objectstorage.test::ReadTestConsistentRegionPeriodicStaticNameBinaryIAMComp", {'IAMApiKey':self.iam_api_key, 'IAMServiceInstanceId':self.service_instance_id, 'objectStorageURI':self.uri_cos}, 1, 'feature/consistent.region.test', True, 90)
-        # operatorDriven
-        self._build_launch_validate("test_read_object_consistent_region_static_name_binary_operatorDriven_iam", "com.ibm.streamsx.objectstorage.test::ReadTestConsistentRegionOperatorDrivenStaticNameBinaryIAMComp", {'IAMApiKey':self.iam_api_key, 'IAMServiceInstanceId':self.service_instance_id, 'objectStorageURI':self.uri_cos}, 1, 'feature/consistent.region.test', True, 90)
-
-    # CONSISTENT REGION: S3ObjectStorageSource, static name, binary file (read 100 blocks of 1 MB block size), no crash (no reset)
-    @unittest.skipIf(th.cos_credentials() == False, "Missing "+th.COS_CREDENTIALS()+" environment variable.")
-    def test_read_object_consistent_region_static_name_binary(self):
-        th.generate_large_bin_file("input.bin") # 100 MB
-        s3.uploadObject(self.s3_client, self.bucket_name, "input.bin", "input.bin")
-        # periodic
-        self._build_launch_validate("test_read_object_consistent_region_static_name_binary_periodic", "com.ibm.streamsx.objectstorage.test::ReadTestConsistentRegionPeriodicStaticNameBinaryComp", {'accessKeyID':self.access_key, 'secretAccessKey':self.secret_access_key, 'bucket':self.bucket_name}, 1, 'feature/consistent.region.test', True, 90)
-        # operatorDriven
-        self._build_launch_validate("test_read_object_consistent_region_static_name_binary_operatorDriven", "com.ibm.streamsx.objectstorage.test::ReadTestConsistentRegionOperatorDrivenStaticNameBinaryComp", {'accessKeyID':self.access_key, 'secretAccessKey':self.secret_access_key, 'bucket':self.bucket_name}, 1, 'feature/consistent.region.test', True, 90)
-
-    # -------------------
-
-    # CONSISTENT REGION: ObjectStorageScan (IAM), operatorDriven, no crash (no reset)
-    @unittest.skipIf(th.iam_credentials() == False, "Missing "+th.COS_IAM_CREDENTIALS()+" environment variable.")
-    def test_scan_consistent_region_operatorDriven_iam(self):
-        s3.uploadObject(self.s3_client_iam, self.bucket_name_iam, "feature/consistent.region.test/etc/input.txt", "scanTestData/input1.txt")
-        s3.uploadObject(self.s3_client_iam, self.bucket_name_iam, "feature/consistent.region.test/etc/input.txt", "scanTestData/input2.txt")
-        s3.uploadObject(self.s3_client_iam, self.bucket_name_iam, "feature/consistent.region.test/etc/input.txt", "scanTestData/input3.txt")
-        self._build_launch_validate("test_scan_consistent_region_operatorDriven_iam", "com.ibm.streamsx.objectstorage.test::ScanTestConsistentRegionOperatorDrivenIAMComp", {'IAMApiKey':self.iam_api_key, 'IAMServiceInstanceId':self.service_instance_id, 'objectStorageURI':self.uri_cos}, 3, 'feature/consistent.region.test')
-
-    # CONSISTENT REGION: ObjectStorageScan (IAM), periodic, no crash (no reset)
-    @unittest.skipIf(th.iam_credentials() == False, "Missing "+th.COS_IAM_CREDENTIALS()+" environment variable.")
-    def test_scan_consistent_region_periodic_iam(self):
-        s3.uploadObject(self.s3_client_iam, self.bucket_name_iam, "feature/consistent.region.test/etc/input.txt", "scanTestData/input1.txt")
-        s3.uploadObject(self.s3_client_iam, self.bucket_name_iam, "feature/consistent.region.test/etc/input.txt", "scanTestData/input2.txt")
-        s3.uploadObject(self.s3_client_iam, self.bucket_name_iam, "feature/consistent.region.test/etc/input.txt", "scanTestData/input3.txt")
-        self._build_launch_validate("test_scan_consistent_region_periodic_iam", "com.ibm.streamsx.objectstorage.test::ScanTestConsistentRegionPeriodicIAMComp", {'IAMApiKey':self.iam_api_key, 'IAMServiceInstanceId':self.service_instance_id, 'objectStorageURI':self.uri_cos}, 3, 'feature/consistent.region.test')
-
-    # -------------------
-
-    # CONSISTENT REGION: ObjectStorageSink (IAM), periodic, no crash (no reset)
-    @unittest.skipIf(th.iam_credentials() == False, "Missing "+th.COS_IAM_CREDENTIALS()+" environment variable.")
-    def test_sink_consistent_region_periodic_iam(self):
-        self._build_launch_validate("test_sink_consistent_region_periodic_iam", "com.ibm.streamsx.objectstorage.test::SinkTestConsistentRegionIAMComp", {'IAMApiKey':self.iam_api_key, 'IAMServiceInstanceId':self.service_instance_id, 'objectStorageURI':self.uri_cos}, 1, 'feature/consistent.region.test', False)
-
-    # -------------------
-
-    # CONSISTENT REGION: ObjectStorageSink PARQUET (IAM), periodic, no crash (no reset)
-    @unittest.skipIf(th.iam_credentials() == False, "Missing "+th.COS_IAM_CREDENTIALS()+" environment variable.")
-    def test_sink_consistent_region_periodic_parquet_iam(self):
-        self._build_launch_validate("test_sink_consistent_region_periodic_parquet_iam", "com.ibm.streamsx.objectstorage.test::ObjectStorageSink_consistent_region_parquetIAMComp", {'IAMApiKey':self.iam_api_key, 'IAMServiceInstanceId':self.service_instance_id, 'objectStorageURI':self.uri_cos, 'drainPeriod':3.0, 'uploadWorkersNum':10}, 1, 'feature/consistent.region.test', False, 120)
-        s3.listObjectsWithSize(self.s3_client_iam, self.bucket_name_iam)
-
 class TestInstall(TestDistributed):
     """ Test invocations of composite operators in local Streams instance using installed toolkit """
